Safari-609.1.11.1/LayoutTests/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 14: InvalidStateError: Document is not fully active
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS
